refactor(frontend): Move project config model to shared module
