+++ frontend/src/app/shared/model/project-config.ts
