+++ frontend/src/app/shared/model/project-zip-request.ts
@@ -1,6 +1,6 @@
 import {ProjectConfig} from './project-config';
 
-export default class ProjectZipRequest extends ProjectConfig {
+export class ProjectZipRequest extends ProjectConfig {
   privacy: string;
   scenarioText: string;
 }
